refactor(routes): import createServerClient from supabase.server
Point auth and settings routes at the server-only Supabase module.
