+++ app/routes/auth.callback.tsx
@@ -1,5 +1,5 @@
 import { LoaderFunctionArgs, redirect } from "@remix-run/node";
-import { createServerClient } from "~/lib/supabase";
+import { createServerClient } from "~/lib/supabase.server";
 import { withAuthHeaders } from "~/lib/auth.server";
 import { getSafeRedirectUrl, isAllowedAuthRedirect } from "~/lib/url-validator";
